+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] – 2025-01-06
+
 ### Added
 - Endpoint for downloading workset data and metadata, optionally with filters applied. [#120](https://github.com/htrc/torchlite-backend/issues/120)
 
@@ -24,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - HTTPX client for mTLS connections to the registry to retrieve worksets. [#125](https://github.com/htrc/torchlite-backend/issues/125)
 
 [unreleased]: https://github.com/htrc/torchlite-backend/compare/0.1.0...HEAD
-[0.1.0]: https://github.com/htrc/torchlite-backend/releases/tag/0.1.0
+[0.1.0]: https://github.com/htrc/torchlite-backend/compare/0.2.0...0.1.0
+[0.2.0]: https://github.com/htrc/torchlite-backend/releases/tag/0.2.0